REF: implement test_astype files
